+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `DiscountBadge`.
 - Snapshot tests for `CategoriesHighlights`.
 - Snapshot tests for `GradientCollapse`.
+- Snapshot tests for `Greeting`.
 
 ## [3.17.2] - 2019-02-18